+++ CHANGELOG.md
@@ -6,7 +6,7 @@
 
 * Added, when an object is selected its material color changes to orange and slightly emissive yellow.
 * Added `Toolbar.vue` featuring existing transforms and views shortcuts into user-friendly buttons.
-* Added command to save views, select them and download a screenshot as PNG.
+* Added command to save views, select them and download a screenshot as image.
 
 ### Changed